case: Move the case's strap wings up
The case's strap wings are currently too low - the thickness of the
straps will cause the case to lift up, while we want to keep pressure on
the arm so that light doesn't spill over from the LEDs into the
photodiodes.

This commit moves the strap up above the vents, and shortens the wings
so that they don't block the SPI connector hole.